adds anybody can edit scorecard
